rename modules to `agent`
